Update bonsai for sketch library changes.
